+++ branches/core_bitmaps/src/srun/srun_job.c
@@ -76,8 +76,8 @@ typedef struct allocation_info {
 	uint32_t                stepid;
 	char                   *nodelist;
 	uint32_t                nnodes;
-	uint16_t                num_cpu_groups;
-	uint32_t               *cpus_per_node;
+	uint32_t                num_cpu_groups;
+	uint16_t               *cpus_per_node;
 	uint32_t               *cpu_count_reps;
 	select_jobinfo_t select_jobinfo;
 } allocation_info_t;
@@ -101,7 +101,7 @@ job_create_noalloc(void)
 {
 	srun_job_t *job = NULL;
 	allocation_info_t *ai = xmalloc(sizeof(*ai));
-	uint32_t cpn = 1;
+	uint16_t cpn = 1;
 	hostlist_t  hl = hostlist_create(opt.nodelist);
 
 	if (!hl) {
